feat: update CORS Proxy route and move welcome message to new route
This commit handles the change in redirection in the Cors-Proxy route when the targetUrl is not provided. Previously, it would return an HTML stint, but now it simply redirects to the base URL ('/'). To complement this change, the welcome message was moved to a separate "greeting" route. This results in the CORS proxy service now being served at '/cors' instead of at the base URL. This has also prompted an update in README.md, changing the URL example to reflect the new path to the CORS Proxy service. The dependency "hono" was also updated to latest available version "3.7.4".